edit from code review

+++ src/main/java/calculator_808/Calculator.java
@@ -2,5 +2,5 @@ package calculator_808;
 
 @FunctionalInterface
 public interface Calculator {
-    int calc(int value1, int value2);
+    long calc(int value1, int value2);
 }
